+++ nuxt/cypress/e2e/accountUsernameSettings.cy.ts
@@ -10,15 +10,15 @@ describe('account username settings page', () => {
     //     expect(resp.redirectedToUrl).to.equal(undefined)
     //   })
     // })
-
-    it('does not find page if user does not exists', () => {
-      cy.request({
-        url: '/account/username-that-does-not-exist/settings',
-        failOnStatusCode: false,
-      }).then((resp) => {
-        expect(resp.status).to.equal(404)
-        expect(resp.redirectedToUrl).to.equal(undefined)
-      })
-    })
+    //
+    // it('does not find page if user does not exists', () => {
+    //   cy.request({
+    //     url: '/account/username-that-does-not-exist/settings',
+    //     failOnStatusCode: false,
+    //   }).then((resp) => {
+    //     expect(resp.status).to.equal(404)
+    //     expect(resp.redirectedToUrl).to.equal(undefined)
+    //   })
+    // })
   })
 })