+++ nuxt/dev.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/bin/bash
 
 set -e
 
@@ -26,20 +26,14 @@ if [ "$CERTIFICATE_PATH" != "" ]; then
   sslCert="$CERTIFICATE_PATH.crt"
   sslKey="$CERTIFICATE_PATH.key"
 else
-  if [ ! -d ".nuxt" ]; then
-    mkdir ".nuxt"
+  if [ ! -d "$THIS/.nuxt" ]; then
+    mkdir "$THIS/.nuxt"
   fi
 
-  if ls "$THIS"/.nuxt/*.key 1> /dev/null 2>&1; then
-    rm "$THIS"/.nuxt/*.key
+  if ! ls "$THIS"/.nuxt/*.key 1> /dev/null 2>&1 || ! ls "$THIS"/.nuxt/*.crt 1> /dev/null 2>&1; then
+    mkcert -install
+    create "localhost" "localhost" "127.0.0.1" "0.0.0.0"
   fi
-
-  if ls "$THIS"/.nuxt/*.crt 1> /dev/null 2>&1; then
-    rm "$THIS"/.nuxt/*.crt
-  fi
-
-  mkcert -install
-  create "localhost" "localhost" "127.0.0.1" "0.0.0.0"
 
   sslCert="$THIS/.nuxt/localhost.crt"
   sslKey="$THIS/.nuxt/localhost.key"