+++ nuxt/pages/event/[username]/[event_name]/invitation/index.vue
@@ -3,9 +3,17 @@
     <div v-if="event" class="flex flex-col gap-4">
       <Breadcrumbs
         :prefixes="[
-          { name: t('events'), to: '../../..', isToRelative: true },
-          { name: routeParamUsername, to: '../..', isToRelative: true },
-          { name: routeParamEventName, to: '..', isToRelative: true },
+          { name: t('events'), to: localePath('/event') },
+          {
+            name: routeParamUsername,
+            to: localePath(`/event/${route.params.username}`),
+          },
+          {
+            name: routeParamEventName,
+            to: localePath(
+              `/event/${route.params.username}/${route.params.event_name}`
+            ),
+          },
         ]"
       >
         {{ t('invitations') }}
@@ -29,6 +37,7 @@ import { useMaevsiStore } from '~/store'
 definePageMeta({
   async validate(route) {
     const { $urql } = useNuxtApp()
+    const store = useMaevsiStore()
 
     const eventIsExisting = await $urql.value
       .query(EVENT_IS_EXISTING_QUERY, {
@@ -37,22 +46,24 @@ definePageMeta({
       })
       .toPromise()
 
-    return !eventIsExisting.error && !!eventIsExisting.data?.eventIsExisting
-  },
-  middleware: [
-    function (_to: any, _from: any) {
-      const route = useRoute()
-      const store = useMaevsiStore()
+    if (eventIsExisting.error) {
+      throw createError(eventIsExisting.error)
+    }
 
-      if (route.params.username !== store.signedInUsername) {
-        throw createError({ statusCode: 403 })
-      }
-    },
-  ],
+    if (!eventIsExisting.data?.eventIsExisting) {
+      return abortNavigation({ statusCode: 404 })
+    }
+
+    if (route.params.username !== store.signedInUsername) {
+      return abortNavigation({ statusCode: 403 })
+    }
+
+    return true
+  },
 })
 
 const route = useRoute()
-const { t } = useI18n()
+const { t, localePath } = useI18n()
 const store = useMaevsiStore()
 
 // queries